chore:#01 tranparent制御をしている。まずはtex1を単独で制御はできた。7
