+++ src/index.html
@@ -22,6 +22,6 @@
     </div>
     <div class="item">Item 3</div>
     <img id="target-img" src="./img/output0.jpg">
-    <script type="module" src="./noise-0/app.js"></script>
+    <script type="module" src="./transparent/app.js"></script>
   </body>
 </html>
