+++ src/transparent/vertex.glsl
